backend/requirements.txt
asgiref==3.8.1
Django==5.0.14
django-cors-headers==4.7.0
djangorestframework==3.16.0
django-filter==24.2
sqlparse==0.5.3
tzdata==2025.2
pandas==2.2.2
coverage==7.4.4
